Move git file to common for sourcemaps
